% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/standby.R
\name{loaders}
\alias{loaders}
\title{CSS Loaders}
\usage{
loaders(uiOutput, type = "default", style = NULL)
}
\arguments{
\item{uiOutput}{An output element to be wrapped within a loader.}

\item{type}{The type of loader to use. heck out \url{https://css-loader.raphaelfabeni.com/} to see the   #' different options.
\itemize{
\item default
\item bar
\item bar-ping-pong
\item border
\item double
\item clock
\item curtain
\item pokeball
\item ball
\item smartphone
\item bouncing
\item music
}}

\item{style}{Custom styling for the loaders.}
}
\description{
Simple CSS loaders
}
\examples{
if (interactive()) {
  library(shiny)

  shinyApp(
    ui = fluidPage(
      useSpinners(),
      actionButton("render", "Render"),
      spinners(plotOutput("plot"))
    ),
    server = function(input, output) {
      output$plot <- renderPlot({
        input$render
        Sys.sleep(3)
        hist(mtcars$mpg)
      })
    }
  )
}

}
