% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preview.R
\name{preview}
\alias{preview}
\alias{previewAlerts}
\alias{previewSpinners}
\title{Preview Alerts}
\usage{
previewAlerts(type = "toast")

previewSpinners(spinner = "threedots")
}
\arguments{
\item{type}{Type of alert/notification. Valid values are:
\itemize{
\item alertify
\item bootbox
\item notice
\item notify
\item tingle
\item toast
}}

\item{spinner}{Type of spinner. The following spinners are available:
\itemize{
\item threedots
\item spinkit
\item vizload
\item spinners
\item loaders
}}
}
\value{
None

None
}
\description{
Preview different types of alerts/notifications.

Preview different types of spinners/loaders.
}
