% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preview.R
\name{previewBuzz}
\alias{previewBuzz}
\title{Preview Alerts}
\usage{
previewBuzz(type = "toast")
}
\arguments{
\item{type}{Type of alert/notification. Valid values are:
\itemize{
\item alertify
\item bootbox
\item notice
\item notify
\item tingle
\item toast
}}
}
\value{
None
}
\description{
Preview different types of alerts/notifications.
}
