% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preview.R
\name{previewSpinners}
\alias{previewSpinners}
\title{Preview Spinners}
\usage{
previewSpinners(spinner = "threeDots")
}
\arguments{
\item{spinner}{Type of spinner. The following spinners are available:
\itemize{
\item threeDots
\item spinkit
\item vizLoad
\item spinners
\item loaders
}}
}
\description{
Preview different types of spinners/loaders.
}
