% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preview.R
\name{previewStandby}
\alias{previewStandby}
\title{Preview Spinners}
\usage{
previewStandby(spinner = "threeDots")
}
\arguments{
\item{spinner}{Type of spinner. The following spinners are available:
\itemize{
\item threedots
\item spinkit
\item vizload
\item spinners
\item loaders
}}
}
\value{
None
}
\description{
Preview different types of spinners/loaders.
}
