% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/standby.R
\name{spinkit}
\alias{spinkit}
\title{SpinKit}
\usage{
spinkit(uiOutput, type = "plane", color = NULL, width = NULL, height = NULL)
}
\arguments{
\item{uiOutput}{An output element to be wrapped within a spinner.}

\item{type}{The type of spinner to use. The following are valid types:
\itemize{
\item plane
\item chase
\item bounce
\item wave
\item pulse
\item flow
\item swing
\item circle
\item circle_fade
\item grid
\item fold
\item wander
}}

\item{color}{The color of the spinner.}

\item{width}{The width of the spinner.}

\item{height}{The height of the spinner.}
}
\description{
Simple CSS spinners.
}
\examples{
if (interactive()) {
  library(shiny)

  shinyApp(
    ui = fluidPage(
      useSpinkit(),
      actionButton("render", "Render"),
      spinkit(plotOutput("plot"))
    ),
    server = function(input, output) {
      output$plot <- renderPlot({
        input$render
        Sys.sleep(3)
        hist(mtcars$mpg)
      })
    }
  )
}

}
