% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/standby.R
\name{spinners}
\alias{spinners}
\title{Single Element CSS Spinners}
\usage{
spinners(uiOutput, type = "1", color = "#0275d8")
}
\arguments{
\item{uiOutput}{An output element to be wrapped within a spinner.}

\item{type}{The type of spinner to use. Any integer between 1 and 8 is valid.}

\item{color}{The color of the spinner.}
}
\description{
A collection of loading spinners animated with CSS
}
\examples{
if (interactive()) {
  library(shiny)

  shinyApp(
    ui = fluidPage(
      useSpinners(),
      actionButton("render", "Render"),
      spinners(plotOutput("plot"))
    ),
    server = function(input, output) {
      output$plot <- renderPlot({
        input$render
        Sys.sleep(3)
        hist(mtcars$mpg)
      })
    }
  )
}

}
