% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deps.R
\name{standbyDeps}
\alias{standbyDeps}
\alias{useThreeDots}
\alias{useSpinkit}
\alias{useVizLoad}
\alias{useSpinners}
\alias{useLoaders}
\title{Include Dependencies}
\usage{
useThreeDots()

useSpinkit()

useVizLoad()

useSpinners()

useLoaders()
}
\description{
Dependencies for spinners/loaders. Include anywhere in the UI of the app.
}
\section{Functions}{

\itemize{
\item {useThreeDots}: Include dependencies for <https://github.com/nzbin/three-dots>
\item {useSpinKit}: Include dependencies for <https://github.com/tobiasahlin/SpinKit>
\item {useVizLoad}: Include dependencies for <https://github.com/RIDICS/Loading-Visualization>
\item {useSpinners}: Include dependencies for <https://github.com/lukehaas/css-loaders> 
\item {useLoaders}: Include dependencies for <https://github.com/raphaelfabeni/css-loader>
}
}

