% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/standby.R
\name{threeDots}
\alias{threeDots}
\title{Three Dots}
\usage{
threeDots(uiOutput, type = "elastic", color = "#9880ff")
}
\arguments{
\item{uiOutput}{An output element to be wrapped within a loader.}

\item{type}{The type of animation to use. Check out \url{https://nzbin.github.io/three-dots/} to see the #' different options.}

\item{color}{The color of the loader.}
}
\description{
CSS loading animation displayed when output is being recalculated.
}
\examples{
if (interactive()) {
  library(shiny)

  shinyApp(
    ui = fluidPage(
      useThreeDots(),
      actionButton("render", "Render"),
      threeDots(plotOutput("plot"))
    ),
    server = function(input, output) {
      output$plot <- renderPlot({
        input$render
        Sys.sleep(3)
        hist(mtcars$mpg)
      })
    }
  )
}

}
