% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/standby.R
\name{vizLoad}
\alias{vizLoad}
\title{Loading Visualization}
\usage{
vizLoad(
  uiOutput,
  type = "bars",
  size = "large",
  color = NULL,
  add_label = FALSE,
  label = "Loading..."
)
}
\arguments{
\item{uiOutput}{An output element to be wrapped within a spinner.}

\item{type}{The type of bar/spinner to use. The following are valid types:
\itemize{
\item bars
\item squares
\item circles
\item dots
\item spinner
\item dashed
\item line
\item bordered-line
}}

\item{size}{The size of the bar/spinner. The following are valid:
\itemize{
\item Large
\item Medium
\item Small
\item Tiny
\item Fluid
}}

\item{color}{The color of the bar/spinner.}

\item{add_label}{Logical; if \code{TRUE}, displays a label below the bar/spinner.}

\item{label}{The label to be displayed below the bar/spinner. \code{add_label} must 
be set to \code{TRUE} to display the label.}
}
\description{
Loading bars and spinners.
}
\examples{
if (interactive()) {
  library(shiny)

  shinyApp(
    ui = fluidPage(
      useVizLoad(),
      actionButton("render", "Render"),
      vizLoad(plotOutput("plot"))
    ),
    server = function(input, output) {
      output$plot <- renderPlot({
        input$render
        Sys.sleep(3)
        hist(mtcars$mpg)
      })
    }
  )
}

}
