% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/standby.R
\docType{package}
\name{standby}
\alias{standby}
\title{\code{standby} package}
\description{
Loading screens for Shiny.
}
\details{
See the README on 
\href{https://github.com/rsquaredacademy/standby}{GitHub}
}
\keyword{internal}
